Fix name generating bug
Currently the fresh names are generated by giving the generator a list
of names that we don't want to crash with, and take whatever the
generator spits out in return. I missed some names on the list, and so
there is a name collision.

I am done with this name generating mechanism. I think we can just
implement an additional counter state in the Elab monad, and use that
instead. It's much, much safer.